+++ django_quickbooks/objects/__init__.py
@@ -1,5 +1,6 @@
 from django.utils.module_loading import import_string
 
+from django_quickbooks.exceptions import QBObjectNotImplemented
 
 ALTERNATIVES = {
     'ItemService': ('Item',),
@@ -7,7 +8,6 @@ ALTERNATIVES = {
 
 
 def import_object_cls(resource_name):
-
     for class_name, alternatives in ALTERNATIVES.items():
         if resource_name in alternatives:
             resource_name = class_name
@@ -15,7 +15,7 @@ def import_object_cls(resource_name):
     try:
         return import_string('%s.%s' % (__package__, resource_name))
     except ImportError:
-        return None
+        raise QBObjectNotImplemented
 
 
 from django_quickbooks.objects.customer import \